CVTK / CI update for analytical test cases

+++ CVTK/test_repro/CVTK_DEBUG_FAST_src/jobcomp_rvtk_datarmor.bash
@@ -42,27 +42,14 @@ echo "OPERATING SYSTEM IS: $OS"
 #
 # compiler options
 #
-if [[ $OS == Linux ]] ; then
+FC=ifort
 
-	LINUX_FC=ifort
 #
-#	set 32 or 64 Bits executable
+# set MPI directories if needed
 #
-	ARCH=`uname -m`
-	echo "PROCESSOR IS: $ARCH"
-	if [[ $ARCH == x86_64 ]] ; then
-		BITS=SIXTYFOUR
-	else
-    	BITS=THIRTYTWO
-	fi
-
-elif [[ $OS == Darwin ]] ; then
-
-        DARWIN_FC=gfortran
-
-else
-  	BITS=THIRTYTWO
-fi
+MPIF90="mpif90"
+MPILIB=""
+MPIINC=""
 
 #
 # set NETCDF directories
@@ -80,13 +67,6 @@ NETCDFLIB=$(nf-config --flibs)
 NETCDFINC=-I$(nf-config --includedir)
 
 #
-# set MPI directories if needed
-#
-MPIF90="mpif90"
-MPILIB=""
-MPIINC=""
-echo $MPIF90
-#
 # set OASIS-MCT (or OASIS3) directories if needed
 #
 PRISM_ROOT_DIR=../../../oasis3-mct/compile_oa3-mct
@@ -99,23 +79,25 @@ XIOS_ROOT_DIR=$HOME/xios-1.0
 # END OF USER'S MODIFICATIONS
 ####################################################
 #
-# Use GNU Make command
+# Use GNU Make command, else make
 #
-MAKE=make
+MAKE=gmake
+which $MAKE > /dev/null 2>&1 || MAKE=make
+
 #
 # clean scratch area
 #
-#####rm -rf $SCRDIR
 mkdir $SCRDIR
+
 #
 # AGRIF sources directory
 #
 AGRIF_SRC=${ROOT_DIR}/AGRIF
+
 #
 # copy SOURCE code
 #
 /bin/cp -f ${SOURCE}/*.F90 $SCRDIR
-/bin/cp -f ${SOURCE}/*.h90 $SCRDIR
 /bin/cp -f ${SOURCE}/*.F   $SCRDIR
 /bin/cp -f ${SOURCE}/*.h   $SCRDIR
 /bin/cp -f ${SOURCE}/Make* $SCRDIR
@@ -127,20 +109,19 @@ AGRIF_SRC=${ROOT_DIR}/AGRIF
 /bin/cp -f ${ROOT_DIR}/PISCES/* $SCRDIR
 /bin/cp -f ${ROOT_DIR}/PISCES/kRGB61* $RUNDIR
 if [[ -e "namelist_pisces" ]] ; then
-	echo "  file namelist_pisces exists in Run directory"
+        echo "  file namelist_pisces exists in Run directory"
 else
-	/bin/cp -f ${SOURCE}/PISCES/namelist_pisces* $RUNDIR
-	echo "  file namelist_pisces copied from source directory"
+        /bin/cp -f ${SOURCE}/PISCES/namelist_pisces* $RUNDIR
+        echo "  file namelist_pisces copied from source directory"
 fi
-#
-# overwrite with local files
-#
-/bin/cp -f *.F90 $SCRDIR
-/bin/cp -f *.h90 $SCRDIR
-/bin/cp -f *.F $SCRDIR
-/bin/cp -f *.h $SCRDIR
-/bin/cp -f Make* $SCRDIR
-/bin/cp -f jobcomp $SCRDIR
+# #
+# # overwrite with local files
+# #
+# /bin/cp -f *.F90 $SCRDIR
+# /bin/cp -f *.F $SCRDIR
+# /bin/cp -f *.h $SCRDIR
+# /bin/cp -f Make* $SCRDIR
+# /bin/cp -f jobcomp $SCRDIR
 #
 #
 # RVTK  files  DEBUG CPP KEYS
@@ -149,7 +130,6 @@ fi
 /bin/cp -f ${SCRDIR}/cppdefs.h.OK ${SCRDIR}/cppdefs.h
 /bin/cp -f ${SCRDIR}/param.h.OK ${SCRDIR}/param.h
 
-#
 # Change directory
 #
 cd $SCRDIR
@@ -161,72 +141,32 @@ CPPFLAGS1="${CROCO_NETCDFINC-$NETCDFINC} -ICROCOFILES/AGRIF_INC"
 #
 # Set compilation options
 #
-if [[ $OS == Linux ]] ; then           # ===== LINUX =====
-	if [[ $LINUX_FC == ifort || $LINUX_FC == ifc ]] ; then
+if [[ $OS == Linux || $OS == Darwin ]] ; then           # ===== LINUX =====
+	if [[ $FC == ifort || $FC == ifc ]] ; then
 		CPP1="cpp -traditional -DLinux -DIfort"
 		CFT1=ifort
-		#FFLAGS1="-O3 -w90 -w95 -cm -72 -fno-alias -i4 -r8 -fp-model precise"
-                # tina suggest adding -C and -debug
-                #FFLAGS1="-O0 -C -g -traceback -72 -fno-alias -i4 -r8 -mcmodel=large -shared-intel -fp-model precise"
-                FFLAGS1="-O1 -72 -fno-alias -i4 -r8 -mcmodel=large -shared-intel -fp-model precise"
-		LDFLAGS1="-Vaxlib $LDFLAGS1"
-        elif [[ $LINUX_FC == gfortran ]] ; then
-		CPP1="cpp -traditional -DLinux"
+		FFLAGS1="-O0 -fno-alias -i4 -r8 -fp-model precise"
+#                FFLAGS1="-O0 -g -i4 -r8 -traceback -check all -check bounds \
+#                       -check uninit -CA -CB -CS -ftrapuv -fpe1"
+		LDFLAGS1="$LDFLAGS1"
+	elif [[ $FC == gfortran ]] ; then
+		CPP1="cpp  -traditional -DLinux"
 		CFT1=gfortran
-		FFLAGS1="-O1 -fdefault-real-8 -fdefault-double-8 -mcmodel=medium"
+		FFLAGS1="-O0 -fdefault-real-8 -fdefault-double-8 "
 #		 FFLAGS1="-O0 -g -fdefault-real-8 -fdefault-double-8 -fbacktrace \
 #			-fbounds-check -finit-real=nan -finit-integer=8888"
 		LDFLAGS1="$LDFLAGS1"
-	fi
-elif [[ $OS == Darwin ]] ; then        # ===== DARWIN =====
-	CPP1="cpp -traditional -DLinux"
-	if [[ $DARWIN_FC == gfortran ]] ; then  
-        	CFT1="gfortran"
-        	FFLAGS1="-O4 -fdefault-real-8 -fdefault-double-8"
-	 	FFLAGS1="-O0 -g -fdefault-real-8 -fdefault-double-8 -fbacktrace \
-			-fbounds-check -finit-real=nan -finit-integer=8888"
-	else
-		CFT1="ifort"
-		FFLAGS1="-O2 -r8 -i4 -g -72"
-#		FFLAGS1="-O0 -g -traceback -debug all -r8 -i4 -g -72"
 	fi
 elif [[ $OS == CYGWIN_NT-10.0 ]] ; then  # ======== CYGWIN =======
         CPP1="cpp -traditional -DLinux"
         CFT1="gfortran"
         FFLAGS1="-O4 -fdefault-real-8 -fdefault-double-8 -march=native -mtune=native"
 elif [[ $OS == AIX ]] ; then           # ===== IBM =====
-	CPP1="/lib/cpp"
+	CPP1="cpp"
 	CFT1="xlf95 -I$HOME/include/"
-	if  [[ $BITS == THIRTYTWO ]] ; then
-		MAKEAGRIF="Makefile.ibm"
-		FFLAGS1="-qfixed -O5 -qstrict -qalias=noaryovrlp -qhot -qrealsize=8 \
-			-qintsize=4 -qarch=auto -qtune=auto -qcache=auto -bmaxdata:0x80000000"
-#		FFLAGS1="-g -qfixed -O2 -qstrict -qalias=noaryovrlp -qrealsize=8 \
-#			-qintsize=4 -qarch=auto -qtune=auto -qcache=auto -bmaxdata:0x80000000"
-	else
-		MAKEAGRIF="Makefile.ibm.64"
-		FFLAGS1="-q64 -qwarn64 -qfixed -qrealsize=8 -qintsize=8 -qhot \
+	MAKEAGRIF="Makefile.ibm.64"
+	FFLAGS1="-q64 -qwarn64 -qfixed -qrealsize=8 -qintsize=8 -qhot \
 			-qalias=noaryovrlp -qthreaded -O3 -qarch=pwr4 -qtune=pwr4 -qunroll=yes"
-	fi
-elif [[ $OS == OSF1 ]] ; then          # ===== COMPAQ =====
-	CPP1="/lib/cpp"
-	CFT1="f95"
-	FFLAGS1="-fast -i4 -r8"
-elif [[ $OS == IRIX64 ]] ; then        # ===== SGI =====
-	CPP1="/usr/freeware/bin/cpp -traditional"
-	CFT1="f90"
-	FFLAGS1="-O2"
-elif [[ $OS == SunOS ]] ; then         # ===== SUN ===== :  tested on SunFire 880 (SPARC III)
-	GREP="/usr/xpg4/bin/grep"      #                         and Sun Ultra-60 (SPARC II)
-	CPP1=/lib/cpp
-	CFT1="f95"
-	if [[ $BITS == THIRTYTWO ]] ; then
-		MAKEAGRIF="Makefile.sun"
-		FFLAGS1="-O5 -xtarget=native -xprefetch -xtypemap=real:64,double:128 -xlibmopt"
-	else
-		MAKEAGRIF="Makefile.sun.64"
-		FFLAGS1="-O5 -xtarget=native64 -xprefetch -xtypemap=real:64,double:128 -xlibmopt "
-	fi
 else
 	echo "Unknown Operating System"
 	exit
@@ -242,8 +182,9 @@ if $($CPP1 testkeys.F | grep -i -q agrifisdefined) ; then
 	FFLAGS1="$FFLAGS1 -IAGRIF"
 	LDFLAGS1="-LAGRIF -lagrif $LDFLAGS1"
 # we use the AGRIF Makedefs.generic definition
-	/bin/cp -f Makedefs.generic.AGRIF Makedefs.generic
+	cp -f Makedefs.generic.AGRIF Makedefs.generic
 fi
+
 #
 # determine if MPI compilation is required
 #
@@ -263,21 +204,7 @@ fi
 #
 FFLAGS1=${CROCO_FFLAGS1-$FFLAGS1}
 CFT1=${CROCO_CFT1-$CFT1}
-#
-# determine if NBQ related solvers (BLAS/LAPACK) are required
-#
-unset COMPILENBQ
-echo "Checking COMPILENBQ..."
-if $($CPP1 testkeys.F | grep -i -q nbqisdefined) ; then
-	echo " => NBQ activated"
-	COMPILENBQ=TRUE
-	#LDFLAGS1="-lblas -llapack $LDFLAGS1"
-	# for datarmor
-	#LDFLAGS1="-mkl=sequential $LDFLAGS1"
-         
-	#
-        FFLAGS1="$FFLAGS1 -ffree-line-length-none"
-fi
+
 #
 # determine if XIOS compilation is required
 #
@@ -291,66 +218,31 @@ if $($CPP1 testkeys.F | grep -i -q xiosisdefined) ; then
         CPPFLAGS1="$CPPFLAGS1 -I$XIOS_ROOT_DIR/inc"
         FFLAGS1="$FFLAGS1 -I$XIOS_ROOT_DIR/inc"
         ln -s $XIOS_ROOT_DIR/bin/xios_server.exe $RUNDIR/.
-        $CPP1 -P -traditional -imacros cppdefs.h  ../field_def.xml_full ../field_def.xml
+        $CPP1 -P -traditional -imacros cppdefs.h  ${ROOT_DIR}/XIOS/field_def.xml_full $RUNDIR/field_def.xml
+        $CPP1 -P -traditional -imacros cppdefs.h  ${ROOT_DIR}/XIOS/domain_def.xml $RUNDIR/domain_def.xml
+        $CPP1 -P -traditional -imacros cppdefs.h  ${ROOT_DIR}/XIOS/iodef.xml $RUNDIR/iodef.xml
 fi
+
 #
 # determine if OASIS compilation is required
 #
 unset COMPILEOASIS
 echo "Checking COMPILEOASIS..."
-PRISM_ROOT_DIR=${PRISM_XIOS_ROOT_DIR-$PRISM_ROOT_DIR}
+PRISM_ROOT_DIR=${CROCO_PRISM_ROOT_DIR-$PRISM_ROOT_DIR}
 if $($CPP1 testkeys.F | grep -i -q oacplisdefined) ; then
     echo " => OASIS activated"
     CHAN=MPI1
-    if $($CPP1 testkeys.F | grep -i -q oacpl_mctisdefined) ; then
-	echo " => OASIS-MCT activated"
-	LIBPSMILE="${PRISM_ROOT_DIR}/lib/libpsmile.${CHAN}.a \
+    LIBPSMILE="${PRISM_ROOT_DIR}/lib/libpsmile.${CHAN}.a \
 		${PRISM_ROOT_DIR}/lib/libmct.a  \
 		${PRISM_ROOT_DIR}/lib/libmpeu.a \
 		${PRISM_ROOT_DIR}/lib/libscrip.a"
-	PSMILE_INCDIR="-I${PRISM_ROOT_DIR}/build/lib/psmile.${CHAN} \
+    PSMILE_INCDIR="-I${PRISM_ROOT_DIR}/build/lib/psmile.${CHAN} \
 		-I${PRISM_ROOT_DIR}/build/lib/mct"
-    elif $($CPP1 testkeys.F | grep -i -q oacpl_oa3isdefined) ; then
-	echo " => OASIS3 activated"
-	LIBPSMILE="${PRISM_ROOT_DIR}/lib/libanaisg.a \
-		${PRISM_ROOT_DIR}/lib/libanaism.a \
-		${PRISM_ROOT_DIR}/lib/libclim.${CHAN}.a \
-		${PRISM_ROOT_DIR}/lib/libpsmile.${CHAN}.a \
-		${PRISM_ROOT_DIR}/lib/libfscint.a  \
-		${PRISM_ROOT_DIR}/lib/libmpp_io.a \
-		${PRISM_ROOT_DIR}/lib/libscrip.a"
-	PSMILE_INCDIR="-I${PRISM_ROOT_DIR}/build/lib/psmile.${CHAN} \
-		-I${PRISM_ROOT_DIR}/build/lib/clim.${CHAN} \
-		-I${PRISM_ROOT_DIR}/build/lib/mpp_io"
-    fi
     COMPILEOASIS=TRUE
-    LDFLAGS1="$LDFLAGS1 $LIBPSMILE"
-    CPPFLAGS1="$CPPFLAGS1 ${PSMILE_INCDIR}"
-    FFLAGS1="$FFLAGS1 ${PSMILE_INCDIR}"
+    LDFLAGS1="$LDFLAGS1 $LIBPSMILE $NETCDFLIB"
+    CPPFLAGS1="$CPPFLAGS1 ${PSMILE_INCDIR} $NETCDFINC"
+    FFLAGS1="$FFLAGS1 ${PSMILE_INCDIR} $NETCDFINC"
 fi
-#
-# rewrite Makedefs according to previous flags
-# with MPI flags if needed
-#
-rm -f Makedefs
-echo 's?$(FFLAGS1)?'$FFLAGS1'?g' > flags.tmp
-echo 's?$(LDFLAGS1)?'$LDFLAGS1'?g' >> flags.tmp
-echo 's?$(CPP1)?'$CPP1'?g' >> flags.tmp
-echo 's?$(CFT1)?'$CFT1'?g' >> flags.tmp
-echo 's?$(CPPFLAGS1)?'$CPPFLAGS1'?g' >> flags.tmp
-sed -f flags.tmp Makedefs.generic > Makedefs
-rm -f flags.tmp
-
-#
-# clean scratch
-#
-$MAKE clobber
-
-#
-# compile the precompiling program
-#
-$MAKE mpc
-
 #
 # prepare and compile the library
 #
@@ -368,14 +260,11 @@ if [[ $COMPILEAGRIF ]] ; then
 		ranlib AGRIF/libagrif.a
 	fi
 #
-	$CPP1 amr.in | grep -v -e ! -e '#' -e % -e '*' > amr.scrum
 	mkdir CROCOFILES
+	mkdir -p CROCOFILES/AGRIF_MODELFILES
+	mkdir -p CROCOFILES/AGRIF_INC
+	$CPP1 amr.in | grep -v -e ! -e '#' -e % -e '*' > CROCOFILES/amr.scrum
 	mv AGRIF/conv CROCOFILES/.
-	mv amr.scrum CROCOFILES/.
-	cd CROCOFILES
-	mkdir AGRIF_MODELFILES
-	mkdir AGRIF_INC
-	cd ..
 	for i in *.h *.h90 ; do
 		echo $i
 		cat cppdefs.h $i | cpp -P | grep -v -e ! -e '#' -e % -e '*' > CROCOFILES/$i
@@ -387,34 +276,22 @@ fi
 # determine if OPENMP compilation is needed
 #
 unset COMPILEOMP
+echo "Checking COMPILEOMP..."
 if $($CPP1 testkeys.F | grep -i -q openmp) ; then
 	COMPILEOMP=TRUE
-	if [[ $OS == Linux ]] ; then
-		if [[ $LINUX_FC == gfortran ]] ; then
+	if [[ $OS == Linux || $OS == Darwin ]] ; then 
+		if [[ $FC == gfortran ]] ; then
 			FFLAGS1="$FFLAGS1 -fopenmp"
-		elif [[ $LINUX_FC == ifort || $LINUX_FC == ifc ]] ; then
-			#FFLAGS1="$FFLAGS1 -openmp"
-			FFLAGS1="$FFLAGS1 -qopenmp"
+		elif [[ $FC == ifort || $FC == ifc ]] ; then
+			FFLAGS1="$FFLAGS1 -openmp"
 		else
 			FFLAGS1="$FFLAGS1 -openmp"
 		fi
-	elif [[ $OS == Darwin ]] ; then 
-		if [[ $DARWIN_FC == gfortran ]] ; then 
-			FFLAGS1="$FFLAGS1 -fopenmp"
-    		else
-			FFLAGS1="$FFLAGS1 -openmp"
-		fi
-        elif [[ $OS == CYGWIN_NT-10.0 ]] ; then
-                FFLAGS1=="$FFLAGS1 -fopenmp"
+	elif [[ $OS == CYGWIN_NT-10.0 ]] ; then
+        FFLAGS1=="$FFLAGS1 -fopenmp"
 	elif [[ $OS == AIX ]] ; then
 		FFLAGS1="$FFLAGS1 -qsmp=omp"
 		CFT1="xlf95_r"
-	elif [[ $OS == OSF1   ]] ; then
-		FFLAGS1="$FFLAGS1 -omp"
-	elif [[ $OS == IRIX64 ]] ; then
-		FFLAGS1="$FFLAGS1 -mp"
-	elif [[ $OS == SunOS  ]] ; then 
-		FFLAGS1="$FFLAGS1 -openmp"
 	fi
 fi
 
@@ -435,6 +312,7 @@ rm -f flags.tmp
 # compile croco
 #
 $MAKE depend
-$MAKE 
-mv croco $RUNDIR
+$MAKE
+  
+[ -f croco ] && mv croco $RUNDIR
 #
